fix(memory): read a code link only when rustdoc would, and guard its limits
Answers the second review of #2261. The published schemas still do not
change: the drift test matches the committed snapshot.

- A bracketed code span was always read as a link: [`0, 1`] would publish
  as `0, 1` and [`value@`] as an empty span. rustdoc reads one only when,
  generic arguments aside, it is a single word, and so does the rewrite;
  [`HashMap<K, V>`] still becomes its code span.
- The macro forms m!() and m!{} count as paths.
- The guard skips instance data as the rewrite does, and flags an
  angle-bracketed ](<crate:: target too. The CHANGELOG and the PR say what
  it catches: a link the rewrite recognizes, or an inline link whose target
  starts with crate::, super::, self:: or Self::. They also give the right
  reason a bare [name] stays: the rewrite cannot tell a resolved link from
  prose, not that rustdoc would leave it.
- reharden_tool_input's doc names the rustdoc pass; dependentSchemas and
  dependencies are walked as name-keyed maps.
- schema.rs was over its file budget, and the rewrite grew it from 1 119
  to 1 398 lines. The tree walks harden applies, the rustdoc rewrite and
  the id widening, move to schema_walks.rs, a #[path] child that keeps
  private access; schema.rs is at 1 053 and its baseline records it.

Proof: dropping the one-word rule, the macro suffixes, the guard's
instance-data skip or its angle-bracket targets each turns its test red.
