+++ scripts/file-budgets-baseline.txt
@@ -1,7 +1,7 @@
 crates/velesdb-core/src/simd_native/x86_avx512.rs	1469
 crates/velesdb-memory/src/context.rs	1097
 crates/velesdb-memory/src/extract.rs	1553
-crates/velesdb-memory/src/schema.rs	1119
+crates/velesdb-memory/src/schema.rs	1053
 crates/velesdb-python/src/agent_memory_service.rs	1097
 crates/velesdb-python/src/options.rs	1123
 crates/velesdb-wasm/src/memory_service.rs	1080
